Refactor `SelectPPMeasurementOp` to use `select_switch` for clarity and consistency; removed `Utility` header and its implementation as it is no longer needed.
