+++ mlir/lib/QEC/Utils/CMakeLists.txt
@@ -16,7 +16,6 @@ target_compile_options(libstim PRIVATE -w)
 
 add_library(QECUtils STATIC
     PauliStringWrapper.cpp
-    Utility.cpp
 )
 
 # features required by Stim
